Improved view_items_by_category.py Added decimal encoder Improved responses.py Improved error message in validation.py
