+++ AnswerKingAPI/layers/common/python/utils/dynamodb_helper.py
@@ -1,0 +1,8 @@
+from decimal import Decimal
+import json
+
+class DecimalEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, Decimal):
+            return str(obj)
+        return super().default(obj)
