+++ AnswerKingAPI/layers/common/python/utils/validation.py
@@ -3,7 +3,7 @@ import json
 def get_path_param(event, key):
     value = event.get('pathParameters', {}).get(key, '').strip()
     if not value:
-        raise ValueError(f"Missing {key}")
+        raise ValueError(f"Missing or invalid {key}")
     return value
 
 def parse_body(event):
